Refactor app container
