+++ template/src/routes/useLanguageFromParams/__tests__/useLanguageFromParams.hook.spec.ts
@@ -1,0 +1,46 @@
+import { renderHook } from '@testing-library/react-hooks';
+import { useLanguageFromParams } from '../useLanguageFromParams.hook';
+import { localesActions } from '../../../modules/locales';
+import { DEFAULT_LOCALE, LOCALES } from '../../../i18n';
+import { store as mockStore } from '../../../../fixtures/store';
+
+const render = () => renderHook(() => useLanguageFromParams());
+
+const mockDispatch = jest.fn();
+const mockParams = jest.fn();
+
+jest.mock('react-router-dom', () => ({
+  useParams: () => mockParams(),
+}));
+
+jest.mock('react-redux', () => ({
+  useSelector: (selector: (state: any) => any) => selector(mockStore),
+  useDispatch: () => mockDispatch,
+}));
+
+describe('useLanguageFromParams: Hook', () => {
+  afterEach(() => {
+    mockParams.mockClear();
+  });
+
+  it('should set proper language based on url', () => {
+    mockParams.mockReturnValue({ lang: 'pl' });
+    render();
+    expect(mockDispatch).toHaveBeenCalledWith(localesActions.setLanguage(LOCALES.POLISH));
+  });
+
+  it('should set default language language is not matched', () => {
+    mockParams.mockReturnValue({ lang: null });
+    render();
+    expect(mockDispatch).toHaveBeenCalledWith(localesActions.setLanguage(DEFAULT_LOCALE));
+  });
+
+  it('should update language when url changes', () => {
+    mockParams.mockReturnValue({ lang: 'pl' });
+    render();
+    mockDispatch.mockClear();
+    mockParams.mockReturnValue({ lang: 'en' });
+    render();
+    expect(mockDispatch).toHaveBeenCalledWith(localesActions.setLanguage(LOCALES.ENGLISH));
+  });
+});
