+++ template/src/routes/useLanguageFromParams/index.ts
@@ -1,0 +1,1 @@
+export { useLanguageFromParams } from './useLanguageFromParams.hook';
