+++ template/src/routes/useStartup/index.ts
@@ -1,0 +1,1 @@
+export { useStartup } from './useStartup.hook';
